requirements.txt
python-telegram-bot==20.7
python-dotenv
gspread
pandas
oauth2client
